FIX: moved float conversion into cutoff, only apply when not floating point

+++ nitorch/io/volumes/mapping.py
@@ -502,7 +502,7 @@ class MappedArray(MappedFile):
             present = (Ellipsis,)
 
         # --- cutoff ---
-        dat[present] = volutils.cutoff(dat[present].float(), cutoff, dim)
+        dat[present] = volutils.cutoff(dat[present], cutoff, dim)
 
         # --- cast + rescale ---
         rand = rand and not indtype.is_floating_point
